Add option to trace state of the RNG.
Option -S now allows to trace the state of the RNG before each action.
Untracing with this information provided may be necessary if the solver
calls the RNG to determine which choice of possible API functions to
call and a problem only occurs with one particular choice (i.e., it can
not be reproduced without reproducing the exact state of the RNG).
This option is disabled by default.