Use true/false for special options.
Boolector expects an integer value and CVC4 expects a string
representing a Boolean value for options we treat specially
(incremental, model generation, unsat assumptions).
Previously, we had a function implemented in each solver that returned a
string representing the 'enable' value for these options. This resulted
in solver dependent traces for these options. We now use "true" for that
and push the responsibility of handling this correctly to the Solver
implementation.