dd: Add options to match occurrence of string in stderr and stdout.
Previously, it was only possible to match against the full stderr
and stdout output when delta debugging. This is a problem for stderr
output where line numbers or pointers (a stack trace) is printed, for
example in case of ASAN errors.

This adds two options to allow matching the occurrence of a given string
rather than the whole stderr/stdout output.