Yices: Use type_is_arithmetic() instead of type_is_real().
Yices strictly distinguishes between Int and Real types, and
type_is_real() answers false for type Int (same with term_is_real() for
Int terms). However, Sort::is_real() must consider that Int is a subtype
of Real, and thus return true for Int sorts. Else, we get a problem when
we create a term of expected SortKind SORT_REAL and the solver returns a
term with type Int (e.g., when creating a value "123"). We delegate sort
inference to the solver, hence we additionally add sort Int under
SortKind SORT_REAL, and file this term under SORT_REAL. When we assert
somewhere that the sort of this term of SortKind SORT_REAL is_real(),
this function must return true for Int sorts.