untrace: Refactor to allow returning list of sorts/terms.
Previously, it was only possible to have an action return a single
created object (sort, term). This addresses this limitation. Actions now
have three return modes: NONE, ID and ID_LIST. All actions now return a
vector of ids, which can be empty (mode NONE), contain a single element
(mode ID) or n elements (mode ID_LIST).

This further fixes a problem with ActionGetValue, where return values
were registered in the term database, but not traced, which can result
in the issue that a term that has not been traced is used as operand.
ActionGetValue now returns a list of values, which are registered to the
term database and traced accordingly.